+++ src/anemoi/inference/commands/run.py
@@ -19,7 +19,7 @@ LOG = logging.getLogger(__name__)
 
 
 class RunCmd(Command):
-    """Inspect the contents of a checkpoint file."""
+    """Run inference from a config yaml file."""
 
     need_logging = False
 
@@ -44,8 +44,7 @@ class RunCmd(Command):
 
         input_state = input.create_input_state(date=config.date)
 
-        if config.write_initial_state:
-            output.write_initial_state(input_state)
+        output.write_initial_state(input_state)
 
         for state in runner.run(input_state=input_state, lead_time=config.lead_time):
             for processor in post_processors: