+++ src/anemoi/inference/context.py
@@ -24,6 +24,17 @@ class Context(ABC):
     verbosity = 0
     development_hacks = {}  # For testing purposes, don't use in production
 
+    # Some runners will set these values, which can be queried by Output objects,
+    # but may remain as None
+
+    reference_date = None
+    time_step = None
+    lead_time = None
+    output_frequency = None
+    write_initial_state = True
+
+    ##################################################################
+
     @property
     @abstractmethod
     def checkpoint(self):