+++ src/anemoi/inference/grib/__init__.py
@@ -1,4 +1,4 @@
-# (C) Copyright 2024 Anemoi contributors.
+# (C) Copyright 2025 Anemoi contributors.
 #
 # This software is licensed under the terms of the Apache Licence Version 2.0
 # which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
@@ -6,18 +6,3 @@
 # In applying this licence, ECMWF does not waive the privileges and immunities
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
-
-import logging
-
-from ..output import Output
-from . import output_registry
-
-LOG = logging.getLogger(__name__)
-
-
-@output_registry.register("none")
-class NoneOutput(Output):
-    """_summary_"""
-
-    def write_step(self, state):
-        pass