+++ src/anemoi/inference/grib/encoding.py
@@ -1,4 +1,4 @@
-# (C) Copyright 2024 Anemoi contributors.
+# (C) Copyright 2024-2025 Anemoi contributors.
 #
 # This software is licensed under the terms of the Apache Licence Version 2.0
 # which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
@@ -9,6 +9,7 @@
 
 
 import logging
+import re
 
 LOG = logging.getLogger(__name__)
 
@@ -16,6 +17,22 @@ LOG = logging.getLogger(__name__)
 GRIB1_ONLY = []
 
 GRIB2_ONLY = ["typeOfGeneratingProcess"]
+
+
+ORDERING = (
+    "edition",
+    "typeOfLevel",
+    "stepType",
+    "productDefinitionTemplateNumber",
+    "eps",
+    "number",
+)
+
+ORDERING = {k: i for i, k in enumerate(ORDERING)}
+
+
+def _ordering(item):
+    return ORDERING.get(item[0], 999)
 
 
 def _param(param):
@@ -30,15 +47,71 @@ def _param(param):
             return "shortName"
 
 
+def _step_in_hours(step):
+    step = step.total_seconds() / 3600
+    assert int(step) == step
+    return int(step)
+
+
+STEP_TYPE = {
+    "accumulation": "accum",
+    "average": "avg",
+    "maximum": "max",
+    "minimum": "min",
+    "instantaneous": None,
+}
+
+
+def encode_time_processing(*, result, template, variable, step, previous_step, edition, ensemble):
+    assert edition in (1, 2)
+
+    if variable.time_processing is None:
+        result["step"] = _step_in_hours(step)
+        # result["startStep"] = _step_in_hours(step)
+        # result["endStep"] = _step_in_hours(step)
+        result["stepType"] = "instant"
+        return
+
+    if previous_step is None:
+        if not variable.is_accumulation:
+            LOG.warning(f"No previous step available for time processing `{variable.time_processing}` for `{variable}`")
+        previous_step = step
+
+    start = _step_in_hours(previous_step)
+    end = _step_in_hours(step)
+
+    result["startStep"] = start
+    result["endStep"] = end
+    result["stepType"] = STEP_TYPE[variable.time_processing]
+
+    if edition == 1:
+        return
+
+    if ensemble:
+        result["productDefinitionTemplateNumber"] = 11
+    else:
+        result["productDefinitionTemplateNumber"] = 8
+
+
+LEVTYPES = {
+    "pl": "isobaricInhPa",
+    "ml": "hybrid",
+    "pt": "theta",
+    "pv": "potentialVorticity",
+}
+
+
 def grib_keys(
     *,
     values,
     template,
-    accumulation,
+    variable,
+    ensemble,
     param,
     date,
     time,
     step,
+    previous_step,
     keys,
     quiet,
     grib1_keys={},
@@ -49,19 +122,6 @@ def grib_keys(
     edition = keys.get("edition")
     if edition is None and template is not None:
         edition = template.metadata("edition")
-        # centre = template.metadata("centre")
-        if edition == 2:
-            productDefinitionTemplateNumber = template.metadata("productDefinitionTemplateNumber")
-            if productDefinitionTemplateNumber in (8, 11) and not accumulation:
-                if f"{param}-accumulation" not in quiet:
-                    LOG.warning(
-                        "%s: Template %s is accumulation but `accumulation` was not specified",
-                        param,
-                        productDefinitionTemplateNumber,
-                    )
-                    LOG.warning("%s: Setting `accumulation` to `True`", param)
-                    quiet.add(f"{param}-accumulation")
-                accumulation = True
 
     if edition is None:
         edition = 1
@@ -75,6 +135,8 @@ def grib_keys(
             result.pop(k, None)
 
     result["edition"] = edition
+
+    result["eps"] = 1 if ensemble else 0
 
     if param is not None:
         result.setdefault(_param(param), param)
@@ -91,45 +153,34 @@ def grib_keys(
         # For organisations that do not use type
         result.setdefault("dataType", result.pop("type"))
 
-    # if stream is not None:
-    #     result.setdefault("stream", stream)
+    result["date"] = date
+    result["time"] = time
 
-    if date is not None:
-        result["date"] = date
+    encode_time_processing(
+        result=result,
+        template=template,
+        variable=variable,
+        step=step,
+        previous_step=previous_step,
+        edition=edition,
+        ensemble=ensemble,
+    )
 
-    if time is not None:
-        result.setdefault("time", time)
+    for k, v in variable.grib_keys.items():
+        if k not in ("domain", "type", "stream", "expver", "class", "param", "number", "step", "date", "time"):
+            if k == "levtype":
+                v = LEVTYPES.get(v)
+                if v is None:
+                    continue
+                k = "typeOfLevel"
+            result.setdefault(k, v)
 
-    # 0: instantaneous, deterministic
-    # 1: instantaneous, ensemble
-    # 8: time processed, deterministic
-    # 11: time processed, ensemble
-
-    if accumulation:
-        if edition == 1:
-            result["step"] = step
-        else:
-            result["startStep"] = 0
-            result["endStep"] = step
-            result["stepType"] = "accum"
-
-        if edition == 2:
-            result["typeOfStatisticalProcessing"] = 1
-            result["productDefinitionTemplateNumber"] = 8
-            if result.get("type") in ("pf", "cf"):
-                result["productDefinitionTemplateNumber"] = 11
-
-    else:
-        result["step"] = step
-        if edition == 2:
-            result["productDefinitionTemplateNumber"] = 0
-            if result.get("type") in ("pf", "cf"):
-                result["productDefinitionTemplateNumber"] = 1
+    result = {k: v for k, v in sorted(result.items(), key=_ordering) if v is not None}
 
     return result
 
 
-def check_encoding(handle, keys):
+def check_encoding(handle, keys, first=True):
     def same(w, v, k):
         if type(v) is type(w):
             return v == w
@@ -165,4 +216,118 @@ def check_encoding(handle, keys):
             mismatches[k] = 'Expected "{}" but got "{}"'.format(v, w)
 
     if mismatches:
+
+        if first:
+            import eccodes
+            from earthkit.data.readers.grib.codes import GribCodesHandle
+
+            handle = GribCodesHandle(eccodes.codes_clone(handle._handle), None, None)
+            return check_encoding(handle, keys, first=False)
+
         raise ValueError(f"GRIB field could not be encoded. Mismatches={mismatches}")
+
+
+def encode_message(*, values, template, metadata, check_nans=False, missing_value=9999):
+    metadata = metadata.copy()  # avoid modifying the original metadata
+    handle = template.handle.clone()
+
+    if check_nans and values is not None:
+        import numpy as np
+
+        if np.isnan(values).any():
+            # missing_value = np.finfo(values.dtype).max
+            missing_value = missing_value
+            values = np.nan_to_num(values, nan=missing_value)
+            metadata["missingValue"] = missing_value
+            metadata["bitmapPresent"] = 1
+
+    if int(metadata.get("deleteLocalDefinition", 0)):
+        for k in ("class", "type", "stream", "expver", "setLocalDefinition"):
+            metadata.pop(k, None)
+
+    metadata.setdefault("generatingProcessIdentifier", 255)
+
+    LOG.debug("GribOutput.metadata %s", metadata)
+
+    single = {}
+    multiple = {}
+    for k, v in metadata.items():
+        if isinstance(v, (int, float, str, bool)):
+            single[k] = v
+        else:
+            multiple[k] = v
+
+    try:
+        # Try to set all metadata at once
+        # This is needed when we set multiple keys that are interdependent
+        handle.set_multiple(single)
+    except Exception as e:
+        LOG.error("Failed to set metadata at once: %s", e)
+        # Try again, but one by one
+        for k, v in single.items():
+            handle.set(k, v)
+
+    for k, v in multiple.items():
+        handle.set(k, v)
+
+    if values is not None:
+        handle.set_values(values)
+
+    return handle
+
+
+class GribWriter:
+    """Write GRIB messages to one or more files."""
+
+    def __init__(self, path, split_output=False):
+        self._files = {}
+        self.filename = path
+
+        if split_output:
+            self.split_output = re.findall(r"\{(.*?)\}", self.filename)
+        else:
+            self.split_output = None
+
+    def close(self):
+        for f in self._files.values():
+            f.close()
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_value, trace):
+        self.close()
+
+    def write(
+        self,
+        *,
+        values,
+        template,
+        metadata,
+        check_nans=False,
+        missing_value=9999,
+    ):
+        handle = encode_message(
+            values=values,
+            check_nans=check_nans,
+            metadata=metadata,
+            template=template,
+            missing_value=missing_value,
+        )
+
+        file, path = self.target(handle)
+        handle.write(file)
+
+        return handle, path
+
+    def target(self, handle):
+
+        if self.split_output:
+            path = self.filename.format(**{k: handle.get(k) for k in self.split_output})
+        else:
+            path = self.filename
+
+        if path not in self._files:
+            self._files[path] = open(path, "wb")
+
+        return self._files[path], path