+++ src/anemoi/inference/outputs/gribfile.py
@@ -12,10 +12,10 @@ import json
 import logging
 from collections import defaultdict
 
-import earthkit.data as ekd
 import numpy as np
 
 from ..decorators import main_argument
+from ..grib.encoding import GribWriter
 from ..grib.encoding import check_encoding
 from . import output_registry
 from .grib import GribOutput
@@ -90,7 +90,9 @@ class GribFileOutput(GribOutput):
         grib1_keys=None,
         grib2_keys=None,
         modifiers=None,
-        **kwargs,
+        output_frequency=None,
+        write_initial_state=None,
+        split_output=True,
     ):
         super().__init__(
             context,
@@ -99,9 +101,11 @@ class GribFileOutput(GribOutput):
             grib1_keys=grib1_keys,
             grib2_keys=grib2_keys,
             modifiers=modifiers,
+            output_frequency=output_frequency,
+            write_initial_state=write_initial_state,
         )
         self.path = path
-        self.output = ekd.new_grib_output(self.path, split_output=True, **kwargs)
+        self.output = GribWriter(self.path, split_output=split_output)
         self.archiving = defaultdict(ArchiveCollector)
         self.archive_requests = archive_requests
         self.check_encoding = check_encoding
@@ -128,10 +132,10 @@ class GribFileOutput(GribOutput):
         try:
             self.collect_archive_requests(
                 self.output.write(
-                    message,
+                    values=message,
                     template=template,
+                    metadata=keys,
                     check_nans=self.context.allow_nans,
-                    **keys,
                 ),
                 template,
                 **keys,